Separate st multistorage package and support the new one in WAL-commands to get rid of a cyclic dependency
