+++ cmd/common/st/folder_list.go
@@ -3,7 +3,7 @@ package st
 import (
 	"github.com/spf13/cobra"
 	"github.com/wal-g/tracelog"
-	"github.com/wal-g/wal-g/internal/multistorage"
+	"github.com/wal-g/wal-g/cmd/common/st/multistorage"
 	"github.com/wal-g/wal-g/internal/storagetools"
 	"github.com/wal-g/wal-g/pkg/storages/storage"
 )
