+++ cmd/pg/wal_prefetch.go
@@ -9,7 +9,6 @@ import (
 	"github.com/wal-g/tracelog"
 	"github.com/wal-g/wal-g/internal"
 	"github.com/wal-g/wal-g/internal/databases/postgres"
-	"github.com/wal-g/wal-g/internal/multistorage"
 )
 
 const WalPrefetchShortDescription = `Used for prefetching process forking
@@ -24,16 +23,8 @@ var WalPrefetchCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		reconfigureLoggers()
 
-		folder, err := internal.ConfigureFolder()
-		tracelog.ErrorLogger.FatalOnError(err)
-
-		failover, err := internal.InitFailoverStorages()
-		tracelog.ErrorLogger.FatalOnError(err)
-
-		folderReader, err := multistorage.NewStorageFolderReader(folder, failover)
-		tracelog.ErrorLogger.FatalOnError(err)
-
-		err = postgres.HandleWALPrefetch(folderReader, args[0], args[1])
+		folderReader := GetWalFolderReader()
+		err := postgres.HandleWALPrefetch(folderReader, args[0], args[1])
 		tracelog.ErrorLogger.FatalOnError(err)
 	},
 }
