+++ internal/storagetools/transfer/handler.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/wal-g/tracelog"
-	"github.com/wal-g/wal-g/internal/multistorage"
+	"github.com/wal-g/wal-g/cmd/common/st/multistorage"
 	"github.com/wal-g/wal-g/pkg/storages/storage"
 	"github.com/wal-g/wal-g/utility"
 )
